+++ dbtk/readers/__init__.py
@@ -8,7 +8,7 @@ consistent interface and automatic format detection.
 
 """
 
-from .base import Reader, Clean
+from .base import Reader
 from .utils import get_reader, open_file
 from .csv import CSVReader
 from .data_frame import DataFrameReader
@@ -20,7 +20,7 @@ from .xml import XMLReader, XMLColumn
 
 # Re-export everything
 __all__ = [
-    'Reader', 'Clean', 'get_reader', 'open_file',
+    'Reader', 'get_reader', 'open_file',
     'CSVReader', 'EDIReader', 'FixedColumn', 'FixedReader', 'JSONReader', 'NDJSONReader', 'XLSReader', 'XLSXReader',
     'XMLColumn', 'XMLReader', 'DataFrameReader', 'open_workbook', 'get_sheet_by_index'
 ]